+++ magik-checks/src/main/java/org/stevenlooman/sw/magik/checks/LhsRhsComparatorEqualCheck.java
@@ -18,11 +18,6 @@ public class LhsRhsComparatorEqualCheck extends MagikCheck {
   public static final String CHECK_KEY = "LhsRhsComparatorEqual";
 
   @Override
-  public boolean isTemplatedCheck() {
-    return false;
-  }
-
-  @Override
   public List<AstNodeType> subscribedTo() {
     return Arrays.asList(
         MagikGrammar.OR_EXPRESSION,