+++ magik-checks/src/main/java/org/stevenlooman/sw/magik/checks/MethodDocCheck.java
@@ -20,11 +20,6 @@ public class MethodDocCheck extends MagikCheck {
   public static final String CHECK_KEY = "MethodDoc";
 
   @Override
-  public boolean isTemplatedCheck() {
-    return false;
-  }
-
-  @Override
   public List<AstNodeType> subscribedTo() {
     return Arrays.asList(MagikGrammar.METHOD_DEFINITION);
   }