+++ magik-checks/src/main/java/org/stevenlooman/sw/magik/checks/UndefinedVariableCheck.java
@@ -20,11 +20,6 @@ public class UndefinedVariableCheck extends MagikCheck {
       "Variable '%s' is expected to be declared, but used as a global.";
 
   @Override
-  public boolean isTemplatedCheck() {
-    return false;
-  }
-
-  @Override
   public List<AstNodeType> subscribedTo() {
     return Arrays.asList(
       MagikGrammar.METHOD_DEFINITION,